removing and adding functionalities
we removed priority queue and the priorities link-list and replaced them with queue. We removed all priority sort methods.

we updated updatePriorityMethod and updateRunningStatus so we incorporate them in the queue logic

we made a 2-3 seconds cool down for the special player, he caonnot re-engage unless the cooldown reach zero.